+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Head
 
+- Added: `Range` helper.
 - Added: `Steeler` webcrawler verifier.
 - Added: `Twitter` webcrawler verifier.
+- Chore: refactoring.
 
 # 1.1.0 - 2016-10-24